requirements.txt
openai~=1.35.3
pyyaml~=6.0.1